Adjusted Camera and Jump Behaviour
The camera no longer locks to the player's exact position when they jump. This was done since it's more useful and less disorienting. As for what's considered ground, I think I should take the top of meshes and use them to define how collision works on a given object, then use tags to determine what is and is not considered ground.