Implemented Cinemachine Camera
Implemented a cinemachine camera to the game. You can use priorities and activation/deactivation to switch between cameras. I want to add triggers that allow virtual cameras to change position based on the player's location in the level,

I may not need the follower camera anymore, but I'm not sure if I want to take it out. I also may switch the controls to applying force in all four directions instead of being for rotation.