Moving Platforms, Camera Target Setup - Trouble With Dampening
I set up moving platforms. I need to use rigidbodies for them to work properly, but this causes a brief stop when they get to the endpoint. I think I could work around this, but it is a problem.

I set up an object that copies the player's transform. I think it'd be best if the camera is a child of it, as it streamlines the process. However, keep in mind that having vcams focus on this target causes stuttering if damping is not set to 0 for all positions. I'm not sure if it's what I want to do, since I want the camera to scroll for player jumps.