Replaced Texture for Ball
Turns out Unity has an issue with the edges of the image texture. For some reason, even if the colours go past the UV map it still blends in colours of the background beyond said colours. The only workaround I've found is filling in the whole image with related colours to avoid this. Either way, try to hide the seams so that it's not visible.