Fixed the Dash Controls
I fixed the dash controls for the game, so now they go in the proper directions. As a side note, I need to have the camera snap instantly back to the checkpoint when the player restarts, rather than have it slowly transition. I also need to make sure there's always a checkpoint at the start of the stage.